Refactor UserSection score utils
