+++ trustScoreUI/src/utils/scoresToBarConverter.js
@@ -1,5 +1,4 @@
-function convertRorScoreToBreakdown(score, maxOverall = 100) {
-  console.log('convertRorScoreToBreakdown', score);
+export function convertRorScoreToBreakdown(score, maxOverall = 100) {
   const subscores = {};
 
   const mapping = {
@@ -31,9 +30,66 @@ function convertRorScoreToBreakdown(score, maxOverall = 100) {
       details: details,
     };
   }
-  console.log(subscores);
 
   return subscores;
 }
 
-export default convertRorScoreToBreakdown;
+export function convertBaseScoreBreakdown(score, maxOverall) {
+    const subscores = {};
+
+     // Name (rank vs max_value)
+  if (score.name) {
+    subscores.name = {
+      total: score.name.max_value,
+      details: [
+        { label: (score.name.max_value === score.name.rank) ? "Perfect Match" : "Match" , value: score.name.rank }
+      ],
+    };
+  }
+
+    // Affiliations
+    if (score.affiliations) {
+      subscores.affiliations = {
+        total: score.affiliations.cumulative_rank,
+        details: [
+          { label: "Count", value: score.affiliations.count },
+          { label: "Avg Rank", value: score.affiliations.avg_rank_per_affiliation },
+        ],
+      };
+    }
+
+    // Emails
+    if (score.emails) {
+      subscores.emails = {
+        total: score.emails.cumulative_rank || score.emails.count,
+        details: [
+          { label: "Count", value: score.emails.count },
+          { label: "Cumulative", value: score.emails.cumulative_rank },
+          { label: "Perfect Match", value: score.emails.perfect_match },
+        ],
+      };
+    }
+
+    // Attributes overall
+    if (score.attributes_with_perfect_match !== undefined) {
+      subscores.attributes = {
+        total: score.attributes_with_perfect_match,
+        details: [
+          { label: "Attributes w/ perfect match", value: score.attributes_with_perfect_match },
+        ],
+      };
+    }
+
+    // Global score
+    if (score.cumulative_rank !== undefined) {
+      subscores.overall = {
+        total: maxOverall,
+        details: [
+          { label: "Cumulative Rank", value: score.cumulative_rank },
+          { label: "Difference", value: Math.round(maxOverall - score.cumulative_rank) },
+        ],
+      };
+    }
+
+    return subscores;
+  }
